Move top level fsp pkg code into an `_engine` module
